Added the edit food activity to deal with food edits + removals, expanded the food class for more functionality later, Moved redundant classes

+++ app/src/main/java/group1/fitnessapp/dietTracker/FoodListAdapter.java
@@ -38,7 +38,7 @@ public class FoodListAdapter extends ArrayAdapter<Food>{
 
         foodName.setText(food.get(position).getName());
         foodSubTxt.setText(food.get(position).getSubText());
-        foodCalories.setText(Integer.toString(food.get(position).getCalories()));
+        foodCalories.setText(Double.toString(food.get(position).getCalories()));
 
         return rowView;
     }
